Unbolded text editor theme to fix parens matching
We had a bug where parens matching did not work when anything in the ace
editor was in bold. I removed all the bold designations in the theme to
get parens matching to work. It is possible there is a way to have bold
and correct parens matching, but in the short run parens matching seemed
more important.